`deepseg_gm.py`: Refactor cropping into cropping+padding
This is necessary because we now fix the input size of the ONNX
model to 200x200. Therefore, images smaller than the input size
must be padded (then cropped afterwards).

(Previously, the input channel itself would be varied to match
the size of the image. But, since we're using a fixed ONNX model,
we can no longer do that.)