`deepseg_gm`: Switch from `constant` padding to `symmetric`
The old use of 'constant' zero-padding created an unrealistic image, because there would be an abrupt transition between the centered image and the zeros surrounding it.

Using 'symmetric' padding creates a more realistic image, and caused the model to produce predictions that are MUCH more similar to the corresponding Keras models.

I tried all of numpy's padding methods, and this one (along with 'wrap') produced the most similar results. All other methods resulted in significant deviations from the Keras results.