`cnn_models_3d.py`: Add bugfix for `UpSampling3d` layer
I think this bug slipped through because `nn_architecture_seg_3d` is not actually called in SCT's current source code. Instead, for 3D models, we use a separate `load_trained_model` function that bypasses the model architecture definition, and loads the model straight from the saved model file.

The only reason I discovered this bug in the first place is because I needed to be able to call `nn_architecture_seg_3d` directly in order to change the name of the output layer to "predictions". (See: the 2 commits that follow this one.)