`onnx_model_generation.ipynb`: Use architecture function calls
These changes are what allow us to generate ONNX models with
"predictions" output layer names.

However, there was no documentation on the correct way to call the
`nn_architecture_seg_3d` function, so I had to guess at what the
right parameters needed to be in order to create the correct architecture.

Because of this, extra care should be taken to ensure that the call to
`nn_architecture_seg_3d` actually produces the same ONNX file as the
old call to `load_trained_model`.