Loosen pytorch pinning.
At some point recently torch put up a full pip repo
(pip --extra-index-url) instead of just a list of files
(pip --find-links), just for their CPU versions, so we
can unpin it. It's not something we use directly at all,
it's something ivadomed calls, and tracking it
here meant duplicating (and when we forget to update,
conflicting with) ivadomed's choices.